CryptoRadarBot/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.6",
    "python-telegram-bot==21.0",
    "requests>=2.32.3",
]
